+++ supreme/client/src/commonMain/kotlin/AttestationClient.kt
@@ -1,6 +1,7 @@
 package at.asitplus.attestation.supreme
 
 import at.asitplus.KmmResult
+import at.asitplus.attestation.supreme.AttestationChallenge.Companion.CURRENT_VERSION
 import at.asitplus.catching
 import at.asitplus.signum.indispensable.asn1.Asn1String
 import at.asitplus.signum.indispensable.asn1.KnownOIDs
@@ -38,6 +39,8 @@ class AttestationClient(client: HttpClient) {
      *  * [AttestationChallenge.validUntil] is earlier than the local system clock
      *  * [AttestationChallenge.issuedAt] is later than the local system clock
      *
+     *  This will also fail when a challenge of a newer version was received
+     *
      * The reason for the second constraint is the simple fact that if the back-end's clock lags behind the local system clock
      * (i.e., challenge issuing time is after [Clock.System.now]), certificate chain validation will fail, due to the
      * leaf certificate's `notBefore` being in the future from the back-end's point of view.
@@ -48,9 +51,12 @@ class AttestationClient(client: HttpClient) {
     suspend fun getChallenge(endpoint: Url): KmmResult<AttestationChallenge> = catching {
         client.get(endpoint).body<AttestationChallenge>().also {
             val now = Clock.System.now()
-            if (it.validUntil?.let { it < now } == true || it.issuedAt > now) throw IllegalStateException(
+            if (it.validUntil.let { it < now } || it.issuedAt > now) throw IllegalStateException(
                 "System time off: issuedAt: ${it.issuedAt}, validUntil: ${it.validUntil}, local system time: $now"
             )
+            it.version?.let { ver ->
+                if (ver > CURRENT_VERSION) throw IllegalArgumentException("Received AttestationChallenge version ${it.version} is newer than locally supported version $CURRENT_VERSION")
+            }
         }
     }
 
@@ -64,8 +70,28 @@ class AttestationClient(client: HttpClient) {
             contentType(ContentType.Application.OctetStream)
             setBody(csr.encodeToDer())
         }.body<AttestationResponse>()
-}
 
+    /**
+     * Truly integrated attestation in a single call.
+     * @throws Throwable Various errors can occur irrespective of attestation:
+     * IO, accessing the platform crypto, not authenticating, etc…
+     *
+     * This is literally a shorthand for:
+     * ```
+     * val challenge = getChallenge(fetchChallengeEndpoint).getOrThrow()
+     * val csr = challenge.createAttestationProof(alias).getOrThrow()
+     * return attest(csr, challenge.attestationEndpointUrl)
+     * ```
+     *
+     * Requires the verifier to pack [KeyConstraints] into the conveyed challenge.
+     */
+    @Throws(Throwable::class)
+    suspend fun performAttestationFlow(alias: String, fetchChallengeEndpoint: Url): AttestationResponse {
+        val challenge = getChallenge(fetchChallengeEndpoint).getOrThrow()
+        val csr = challenge.createAttestationProof(alias).getOrThrow()
+        return attest(csr, challenge.attestationEndpointUrl)
+    }
+}
 
 /**
  * Creates a signed CSR from a received [AttestationChallenge] according to [AttestationChallenge.keyConstraints].
@@ -86,8 +112,7 @@ suspend fun AttestationChallenge.createAttestationProof(
     val params = keyConstraints?.algorithmParameters
         ?: throw IllegalArgumentException("No algorithm specified. Refusing to automatically create an attested key")
     val protectionParameters = keyConstraints?.keyProtection
-    val signer = PlatformSigningProvider.createSigningKey(alias) {
-
+    PlatformSigningProvider.createSigningKey(alias) {
         when (params) {
             is KeyConstraints.AlgorithmParameters.EC -> ec {
                 curve = params.curve
@@ -132,6 +157,14 @@ suspend fun AttestationChallenge.createAttestationProof(
             Asn1String.UTF8(getDeviceName()).encodeToTlv()
         )
     ) else listOf()
+    val signer = PlatformSigningProvider.getSignerForKey(alias) {
+        keyConstraints?.keyProtection?.authPrompt?.let {
+            unlockPrompt {
+                message = it.message
+                cancelText = it.cancelText
+            }
+        }
+    }.getOrThrow()
     return signer.createCsr(this, additionalAttributes = additionalAttributes)
 }